break out column and cell parser classes into a dedicated ::Export::Parsers module
