+++ lib/analytics/export.rb
@@ -2,9 +2,8 @@ require_relative "export/model"
 require_relative "export/builder"
 require_relative "export/buildable"
 require_relative "export/csv"
-require_relative "export/column_parser"
-require_relative "export/export_record"
 require_relative "export/message"
+require_relative "export/parsers"
 
 module Analytics
   module Export
