+++ lib/analytics/export/parsers.rb
@@ -1,0 +1,9 @@
+require_relative "parsers/cell"
+require_relative "parsers/column"
+
+module Analytics
+  module Export
+    module Parsers
+    end
+  end
+end
